Update consent resource class for idiomatic REST
