+++ .gitignore
@@ -149,4 +149,3 @@ hs_err_pid*
 
 ### OSX ###
 .DS_STORE
-*.yml
