+++ consent-ws/src/main/java/org/genomebridge/consent/http/resources/ConsentResource.java
@@ -6,17 +6,17 @@ import org.genomebridge.consent.http.service.ConsentAPI;
 import org.genomebridge.consent.http.service.UnknownIdentifierException;
 
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.UriInfo;
+import java.net.URI;
 
-@Path("consent/{id}")
+@Path("consent")
 public class ConsentResource extends Resource {
 
     private ConsentAPI api;
 
-    public ConsentResource() {
-        this.api = AbstractConsentAPI.getInstance();
-    }
-
+    @Path("{id}")
     @GET
     @Produces("application/json")
     public Response describe(@PathParam("id") String id) {
@@ -29,6 +29,19 @@ public class ConsentResource extends Resource {
     }
 
     @POST
+    @Consumes("application/json")
+    public Response createConsent(@Context UriInfo info, Consent rec) {
+        try {
+            Consent consent = api.create(rec);
+            URI uri = info.getRequestUriBuilder().path("{id}").build(consent.consentId);
+            return Response.created(uri).build();
+        } catch (Exception e) {
+            return Response.serverError().entity(e).build();
+        }
+    }
+
+    @Path("{id}")
+    @PUT
     @Consumes("application/json")
     @Produces("application/json")
     public Response update(@PathParam("id") String id, Consent updated) {
@@ -44,6 +57,10 @@ public class ConsentResource extends Resource {
 
     private Consent populateFromApi(String id) throws UnknownIdentifierException {
         return api.retrieve(id);
+    }
+
+    public ConsentResource() {
+        this.api = AbstractConsentAPI.getInstance();
     }
 
 }
