+++ consent-ws/src/test/java/org/genomebridge/consent/http/AssociationTest.java
@@ -353,7 +353,7 @@ public class AssociationTest extends ConsentServiceTest {
     private String setupConsent() {
         Client client = ClientBuilder.newClient();
         Consent rec = new Consent(false, new Everything(), UUID.randomUUID().toString());
-        Response response = checkStatus( CREATED, put(client, consentPath(), rec) );
+        Response response = checkStatus( CREATED, post(client, consentPath(), rec) );
         String createdLocation = checkHeader(response, "Location");
         String consent_id = createdLocation.substring(createdLocation.lastIndexOf("/") + 1);
         System.out.println(String.format("setupConsent created consent with id '%s' at location '%s'", createdLocation, consent_id));
